requirements.txt
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
attrs==23.2.0
backoff==2.2.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.6
defusedxml==0.7.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
frozenlist==1.4.1
fuzzywuzzy==0.18.0
greenlet==3.0.3
idna==3.7
inference-sdk==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
Levenshtein==0.25.1
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
pyserial==3.5
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
PyYAML==6.0.1
rapidfuzz==3.9.0
requests==2.31.0
scipy==1.13.0
six==1.16.0
SQLAlchemy==2.0.30
supervision==0.20.0
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
